requirements.txt
tensorflow>=2.12.0
numpy
opencv-python
scikit-learn
matplotlib
seaborn
tqdm
scipy
imgaug
Pillow
pandas
